Added more debug info in patch.sh

+++ patch.sh
@@ -11,8 +11,22 @@ sudo apt-get install libopenal1 libalut0 libogg0 libvorbis*
 
 cp AstroMenace.patch ../
 cd ..
+
+if [ ! -e AstroMenace.patch ]
+then
+  echo "ERROR: AstroMenace.patch not found in parent folder of AstroMenaceCheat"
+  exit
+fi
+
 patch -p0 -i AstroMenace.patch
 rm AstroMenace.patch
+
+if [ ! -d AstroMenaceCheat ]
+then
+  echo "ERROR: AstroMenaceCheat folder not found in parent folder of AstroMenaceCheat"
+  exit
+fi
+
 cd AstroMenaceCheat
 cmake ./
 make
